requirements.txt
flask
Pillow
tensorflow-cpu
numpy
gunicorn